Users and accounts controller for testing

+++ sources/webapplication/app/Controller/AccountsController.php
@@ -5,21 +5,20 @@ class AccountsController extends AppController {
 
     public $name = 'Accounts';
     public $helpers = array('Html','Js');
-    public $components = array('RequestHandler','Recaptcha.Recaptcha' => array('actions' => array('register'),'error' => 'tratra'));
+    public $components = array('RequestHandler','Recaptcha.Recaptcha' => array('actions' => array('register')));
     public $layout = 'login';
   
   
     public function beforeFilter(){
         parent::beforeFilter();
-        $this->Auth->allow('index','login','loginzalogin','activate','reactivation','confirm_email','confirm_user_data','register');
+        $this->Auth->allow('loginzalogin','activate','reactivate','confirm_email','confirm_user_data');
         if ($this->Auth->user() && 
-            in_array($this->params['action'], array('login',
+            in_array($this->params['action'], array(
                                                     'loginzalogin',
                                                     'activate',
-                                                    'reactivation',
+                                                    'reactivate',
                                                     'confirm_email',
                                                     'confirm_user_data',
-                                                    'register'
                                                 )
                     )) {
 			$this->redirect('/');
@@ -37,13 +36,13 @@ class AccountsController extends AppController {
             $result = $this->Account->getLoginzaUser($this->request->data['token']);
             switch($result['status']){
                 case 'active':{
-                    $this->Auth->login($result['Account']);
+                    $this->Auth->login($result);
                     $this->redirect($this->Auth->redirect()); break;
                 }
                 case 'notActive':{
                     $this->Session->write('tmpUser',$result);
-                    $this->Session->setFlash(__('Your account not active.','alert',array('class'=>'alert-info'))); 
-                    $this->redirect(array('action'=>'reactivation')); break;   
+                    $this->Session->setFlash(__('Your account not active.'),'alert',array('class'=>'alert-info')); 
+                    $this->redirect(array('action'=>'reactivate')); break;   
                 }
                 case 'newUser':{
                     $this->Session->write('tmpUser',$result);
@@ -51,7 +50,7 @@ class AccountsController extends AppController {
                 }
                 case 'error':{
                     $this->Session->setFlash(__($result['msg']),'alert',array('class'=>'alert-error'));
-                    $this->redirect(array('action'=>'login'));
+                    $this->redirect('/');
                 }
                 
              }
@@ -60,54 +59,56 @@ class AccountsController extends AppController {
 		$this->redirect(array('action'=>'login'));
     }
     
-    public function reactivation(){
+    public function reactivate(){
         
         if($this->Session->check('tmpUser')){
             $user_profile = $this->Session->read('tmpUser');
             if ($this->request->is('post') || $this->request->is('put')) {
                 $this->Session->delete('tmpUser');
-                $result = $this->Account->activation($user_profile['Account']['id']);
-                $this->Session->setFlash($result['msg']);
-                $this->redirect(array('action'=>'login'));
+                if($this->Account->reactivate($user_profile[$this->modelClass]['id'],$this->name)){
+	                  $this->Session->setFlash(__('Код активации аккаунта был выслан Вам на почту.'),'alert',array('class'=>'alert-success'));
+                      $this->redirect('/');  
+	            }
+                $this->Session->setFlash(__('Возникли проблемы при отправке КодА активации аккаунта.'),'alert',array('class'=>'alert-success'));
+                $this->redirect('/');
             }else{
                 $this->request->data = $user_profile;    
             }
         }else{
-            $this->redirect(array('action'=>'login'));
+            $this->redirect('/');
         }
     }
     
     public function confirm_email(){
         
          if(!$this->Session->check('tmpUser')){
-                $this->redirect(array('action' => 'login'));
+                $this->redirect('/');
          }
         $user = $this->Session->read('tmpUser');
         if ($this->request->is('post') || $this->request->is('put')) {
             $user['email'] = $this->request->data['User']['email'];
-            $result = $this->Account->User->checkEmail($this->request->data);
-            switch($result['status']){
-                case'error':{
-                    $this->Session->setFlash($result['msg'] ,'alert');
-                    return false;
-                }    
-                case 'new':{
-                    $this->Session->write('tmpUser',$user);
-                    $this->redirect(array('action'=>'confirm_user_data')); break;
-                }
-                case 'already':{
-                    $user['user_id'] = $result['data']['id'];
-                    if($this->Account->save($user)){
-                        $resultActivate = $this->Account->activation($this->Account->getLastInsertID());
-                        $this->Session->setFlash($resultActivate['msg'],'alert',array('class'=>'alert-success'));
-                        $this->redirect(array('action'=>'login'));
-                    }else{
-                        $this->Session->setFlash('Error create account','alert',array('class'=>'alert-error'));
-                        $this->redirect(array('action'=>'login'));
+            if(!$this->Account->User->validateEmail($this->request->data)){
+                return $this->Session->setFlash(__('Возникла ошибка при заполнении. Пожалуйста, попробуйте еще раз.'),'alert',
+                                                    array('class'=>'alert-error'));
+            }
+            if ($id = $this->Account->User->checkEmail($this->request->data)){
+                $this->Session->delete('tmpUser');
+                $user['user_id'] = $id;
+                $user['activate_token'] = $this->Account->User->generateToken();
+                if($this->Account->save($user)){
+                    if($this->Account->sendActivationAccount($this->Account->getLastInsertID(),$this->name)){
+	                  $this->Session->setFlash(__('Код активации аккаунта был выслан Вам на почту.'),'alert',array('class'=>'alert-success'));
+                      $this->redirect('/');  
                     }
-                    $this->Session->delete('tmpUser');
+                    $this->Session->setFlash(__('Вы успешно прошли регистрацию. Возникли проблемы при отправке КодА активации аккаунта.'),'alert',array('class'=>'alert-success'));
+                    $this->redirect('/');
+                }else{
+                    $this->Session->setFlash('Error create account','alert',array('class'=>'alert-error'));
+                    $this->redirect('/');
                 }
             }
+            $this->Session->write('tmpUser',$user);
+            $this->redirect(array('action'=>'confirm_user_data'));
         }else{
            $this->request->data['User']['email'] = $user['email'];
         }
@@ -116,28 +117,31 @@ class AccountsController extends AppController {
     public function confirm_user_data(){
         
          if(!$this->Session->check('tmpUser')){
-            $this->redirect(array('action' => 'login'));
+            $this->redirect('/');
          }
-         $userSocial = $this->Session->read('tmpUser');
+         $userTmp = $this->Session->read('tmpUser');
          if ($this->request->is('post') || $this->request->is('put')) {
+            $this->request->data['User']['password'] = $this->request->data['User']['password_confirm'] = $this->Account->User->generateToken();
             if($this->Account->User->save($this->request->data)){
-                $saveAccount['provider'] = $userSocial['provider'];
-                $saveAccount['uid'] = $userSocial['uid'];
-                $saveAccount['user_id'] = $this->Account->User->getLastInsertID();
-                if($this->Account->save($saveAccount)){
-                    $resultActivate = $this->Account->activation($this->Account->getLastInsertID());
-                    $this->Session->setFlash($resultActivate['msg']);
+                $userTmp['user_id'] = $this->Account->User->getLastInsertID();
+                $userTmp['activate_token'] = $this->Account->User->generateToken();
+                if($this->Account->save($userTmp)){
+                    if($this->Account->sendActivationAccount($this->Account->getLastInsertID(),$this->name)){
+		                  $this->Session->setFlash(__('Код активации аккаунта был выслан Вам на почту.'),'alert',array('class'=>'alert-success'));
+                          $this->redirect('/');  
+		            }
+                    $this->Session->setFlash(__('Вы успешно прошли регистрацию. Возникли проблемы при отправке КодА активации аккаунта.'),'alert',array('class'=>'alert-success'));
+                    $this->redirect('/');
                 }else{
                     $this->Session->setFlash('Error create user profile','alert',array('class'=>'alert-error'));
                 }
-                $this->redirect(array('action'=>'login'));    
                 $this->Session->delete('tmpUser');    
+                $this->redirect('/');    
             }else {
-               $this->Session->setFlash('Error create account','alert',array('class'=>'alert-error'));
-               return false;
+              return $this->Session->setFlash('Error create account','alert',array('class'=>'alert-error'));
             }
          }else{
-                $this->request->data['User'] = $userSocial;
+                $this->request->data['User'] = $userTmp;
          }
     }  
   
@@ -145,11 +149,11 @@ class AccountsController extends AppController {
     	if ($hash){
             if($this->Account->activate($hash)){
                 $this->Session->setFlash('Your account has been activated, please log in.','alert',array('class'=>'alert-success'));
-        		return $this->redirect(array('action' => 'login'));    
+        		$this->redirect('/');    
             }    
        }
        $this->Session->setFlash(__('Invalid key.'));
-       $this->redirect(array('action' => 'login'));
+       $this->redirect('/');
     }
